pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ecg-lab"
version = "0.1.0"
description = "ECG time-series processing utilities, monitoring tools, and training pipelines."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy",
    "pandas",
    "pyarrow",
    "scipy",
]

[project.optional-dependencies]
dev = [
    "pytest",
]
ml = [
    "neurokit2",
    "scikit-learn",
    "torch",
]
monitor = [
    "pyqt5",
    "pyqtgraph",
    "pyserial",
    "rich",
]

[project.scripts]
ecg-lab = "ecg_lab.cli:main"

[tool.setuptools]
packages = ["ecg_core", "ecg_lab"]
